backend/User-service/requirements.txt
flask
psycopg2-binary
gunicorn
blinker==1.9.0
click==8.3.3
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
prometheus_client==0.25.0
psycopg2-binary==2.9.12
Werkzeug==3.1.8